General Commit & Superficial Push
Many changes to meet Task 2.  Note that this is an interim push, with
programs in the middle of being updated / tested / etc.  So, pulling
from this particular codeset is NOT advised.  DougEagleman is the
current release, and should be used as the base codeset for downloads,
etc. until the next release.