pulling data using the platform
The camp_interface file is updated so my function will run on the
platform. The assignLifeStage function is what is called by the
platform.  The getDataForLifeStageAssign function runs the queries and
pulls the data from the database.

This all seems to be working. I need to talk with Trent about the actual
life stage assignment.